# Align Vue implementation with others