--- a/vue/client/index.html
+++ b/vue/client/index.html
@@ -2,17 +2,29 @@
 <html>
   <head>
     <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/base.css" />
     <style>
-    body {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      height: 100vh;
-      background-color: #f0f0f0;
-      margin: 0;
-    }
+      body {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        height: 100vh;
+        background-color: #f0f0f0;
+        margin: 0;
+      }
+      #wrapper {
+        width: 960px;
+        height: 720px;
+        position: relative;
+        background-color: white;
+      }
+      .tile {
+        position: absolute;
+        width: 10px;
+        height: 10px;
+        background-color: #333;
+      }
     </style>
     <!-- head -->
   </head>