@icarusgk 2024-08-27
We could also replace `ref()` usage with `shallowRef()` for a performance improvement.