@icarusgk 2024-08-27
@zsilbi Could you run the other frameworks (with the fixes) and post them, please? I want to make a graph with the results.