@brunnerh 2024-08-28
> Using scoped CSS in the Vue component is unfair, compared to the other implementations.

I somewhat disagree, scoped styles seem idiomatic. Svelte also uses scoped styles but is not particularly impacted by that. If scoped styles cause this much of a performance hit in Vue, that seems like a real issue.

Most of these other frameworks just don't address styling at all, so to me it would make more sense to bring in something *there* to address that. Proponents of the other frameworks probably would not be very happy about that, though, since it would drag their results down a bit and include some additional code/library. (Then again, I don't think they should get to complain if their frameworks are simply incomplete.)