@Rich-Harris 2024-08-28
> I somewhat disagree, scoped styles seem idiomatic

That seems right to me — Svelte would get a tiny boost from having unscoped styles commensurate with the slightly reduced payload size, but we'd never advise people to avoid scoped styles for performance reasons. I definitely wouldn't describe it as 'unfair'. If the slowdown is really that dramatic — almost 3x! — then I would expect to see prominent warnings in the Vue docs, but I'm not aware of any.

Removing scoped styles from the Vue implementation certainly isn't an 'alignment', given that Svelte uses scoped CSS while the Solid, React and fastify-html implementations all use inline styles. If anything, this PR makes the Vue implementation less like the others.

(You _could_ make the same argument about `ref` vs `shallowRef` — while `shallowRef` is totally fine for a benchmark with no fine-grained client-side reactivity, it's arguably less idiomatic — but since the other benchmarks are basically just using normal variables this is less clear-cut.)