@Rich-Harris 2024-08-28
(Actually I mis-spoke — the other implementations _are_ using inline styles, but only for the same reason as the Vue and Svelte ones are, namely the per-div `left` and `top` styles. Nonetheless I think the point stands — it makes sense to compare idiomatic code with idiomatic code, since that's what end users experience.)