@mcollina 2024-08-28
@Rich-Harris so you recommend _not_ to land this?