@zsilbi 2024-08-28
> @zsilbi Could you run the other frameworks (with the fixes) and post them, please? I want to make a graph with the results.

|Name|Avg Req/Sec|
|--|--|
|fastify-html|905.37|
|fastify-htmx|677.3|
|preact|843.3|
|react|522.4|
|solid|1,155.28|
|svelte|921.7|
|vue|128.9|

Using: https://github.com/platformatic/ssr-performance-showdown/commit/da9d95219c0b0cf252ebca7f1e9ff0789eb95560


|Name|Avg Req/Sec|
|--|--|
|vue (optimized)|1,027.82|

Using this PR.