@RomainLanz 2024-08-28
> I agree that the code for all frameworks should be idiomatic. This seems to hyper optimize the Vue code. I've been writing Vue for over 4 years and I would rarely use shallowRef. Also moving the CSS out of the Vue component is super strange and would never be done in an idiomatic Vue codebase

I have been doing Vue for a long time and often use `shallowRef`. It is not hyper-optimized but properly coded.

I don't find any reason not to merge it.