@gautemo 2024-08-28
Maybe to late since it's merged 😄 But neither `ref` or `shallowRef` should be used. `tiles` is not a reactive variable (same as other implementations), so it should be the exact same code as the other implementations. You can just copy the exact same code from e.g. `fastify-html`:

```vue
<script setup>
const wrapperWidth = 960
const wrapperHeight = 720
const cellSize = 10
const centerX = wrapperWidth / 2
const centerY = wrapperHeight / 2

let idCounter = 0
let angle = 0
let radius = 0

const tiles = []
const step = cellSize

let x
let y
while (radius < Math.min(wrapperWidth, wrapperHeight) / 2) {
  x = centerX + Math.cos(angle) * radius
  y = centerY + Math.sin(angle) * radius

  if (x >= 0 && x <= wrapperWidth - cellSize && y >= 0 && y <= wrapperHeight - cellSize) {
    tiles.push({ x, y, id: idCounter++ })
  }

  angle += 0.2
  radius += step * 0.015
}
</script>
```