@zsilbi 2024-08-28
> Maybe to late since it's merged 😄 But neither `ref` or `shallowRef` should be used. `tiles` is not a reactive variable (same as other implementations), so it should be the exact same code as the other implementations. You can just copy the exact same code from e.g. `fastify-html`

I tested this before switching to `shallowRef` and found no improvement at all.