@zsilbi 2024-08-28
I re-ran the original version of benchmark like 50 times and found inconsistency in the performance when the `<style>` block is present.

I restarted the node process between the runs, and 
- 3 out of 10 times the benchmark ran like the `<style>` was not there, around 330 req/sec
- other times I got similar performance like originally, below 140 req/sec

I could also replicate this using `node  --inspect`:

## Normal run

```
┌─────────┬───────┬───────┬───────┬────────┬─────────┬──────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%    │ Avg     │ Stdev    │ Max    │
├─────────┼───────┼───────┼───────┼────────┼─────────┼──────────┼────────┤
│ Latency │ 23 ms │ 29 ms │ 66 ms │ 100 ms │ 31.5 ms │ 12.78 ms │ 153 ms │
└─────────┴───────┴───────┴───────┴────────┴─────────┴──────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬───────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5% │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼───────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 277     │ 277     │ 311     │ 354   │ 311.9   │ 22.25   │ 277     │
├───────────┼─────────┼─────────┼─────────┼───────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 65.7 MB │ 65.7 MB │ 73.8 MB │ 84 MB │ 73.9 MB │ 5.27 MB │ 65.7 MB │
└───────────┴─────────┴─────────┴─────────┴───────┴─────────┴─────────┴─────────┘
```

![310_rps](https://github.com/user-attachments/assets/2fc1a608-cbd7-4e54-9cac-37867ddc17af)
[310_rps.cpuprofile](https://github.com/user-attachments/files/16780861/310_rps.cpuprofile)


### Slow run
```
┌─────────┬───────┬───────┬────────┬────────┬──────────┬───────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5%  │ 99%    │ Avg      │ Stdev     │ Max    │
├─────────┼───────┼───────┼────────┼────────┼──────────┼───────────┼────────┤
│ Latency │ 27 ms │ 31 ms │ 275 ms │ 468 ms │ 83.17 ms │ 100.06 ms │ 809 ms │
└─────────┴───────┴───────┴────────┴────────┴──────────┴───────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬───────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg   │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼───────┼─────────┼─────────┤
│ Req/Sec   │ 39      │ 39      │ 47      │ 315     │ 117.9 │ 114.19  │ 39      │
├───────────┼─────────┼─────────┼─────────┼─────────┼───────┼─────────┼─────────┤
│ Bytes/Sec │ 9.25 MB │ 9.25 MB │ 11.1 MB │ 74.7 MB │ 28 MB │ 27.1 MB │ 9.25 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴───────┴─────────┴─────────┘
```
![120_rps](https://github.com/user-attachments/assets/9c31a4d6-7780-4cac-b44b-3298454f58e3)
[120_rps.cpuprofile](https://github.com/user-attachments/files/16780871/120_rps.cpuprofile)

I found no difference in memory consumption between these runs.

Tests were run on:
WSL2 Ubuntu 22.04.4 LTS
Node v20.17.0 LTS (and checked also on v20.10.0)