@davidmatter 2024-08-28
> But neither `ref` or `shallowRef` should be used. `tiles` is not a reactive variable

That's right. No reactivity needed here.